Move Numeric to Core and remove Utility dependency
