+++ src/Core/CMakeLists.txt
@@ -1,11 +1,12 @@
 set(public_files
-    "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/BitSet.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/BinaryStream.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/BitSet.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/EnumFlags.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/FileStream.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/FileSystem.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/LocoFixedVector.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/MemoryStream.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Numeric.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Prng.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Span.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Stream.hpp"
@@ -16,6 +17,7 @@ set(private_files
     "${CMAKE_CURRENT_SOURCE_DIR}/src/BinaryStream.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/FileStream.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/MemoryStream.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Numeric.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Prng.cpp"
 )
 
@@ -23,6 +25,7 @@ set(test_files
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/EnumFlagsTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/FileStreamTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/MemoryStreamTests.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/tests/NumericTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/PrngTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/TraitsTest.cpp"
 )
@@ -40,6 +43,5 @@ target_link_libraries(Core
     PUBLIC
         nonstd::span-lite
         fmt::fmt
-    PRIVATE
-        Utility)
+)
