+++ src/Core/include/OpenLoco/Core/Numeric.hpp
