+++ src/Core/src/Numeric.cpp
